+++ src/ckanapi_harvesters/ckan_api/ckan_api_params.py
@@ -45,8 +45,10 @@ class CkanApiParamsBasic:
         # limits
         self.default_limit_list:Union[int,None] = 100   # limit the number of entries per list response (used as default value)
         self.default_limit_read:Union[int,None] = 5000  # limit the number of entries per response (used as default value)
+        self.max_requests_attempts:int = 3  # when performing an API action, retry depending on the received HTTP error code
         self.max_requests_count:int = 0  # when automating multiple requests, the number of requests is limited by this parameter. 0 means no limit.
         # timeouts
+        self.time_between_attempts:float = 10  # when a retry was decided, wait this time before a new tentative (* request number)
         self.multi_requests_timeout:float = 0  # when automating multiple requests, the total time elapsed is limited by this parameter (evaluated between each request). 0 means no limit.
         self.multi_requests_time_between_requests:float = 0.100  # when automating multiple requests, wait this additional time (in seconds) between each request
         self.requests_timeout:Union[float,None] = 100  # timeout per request sent to the requests module